Se ajustan DTOs

+++ logistica_transporte_backend/src/main/java/com/logistica/transporte/dto/BodegaDTO.java
@@ -1,10 +1,9 @@
 package com.logistica.transporte.dto;
 
-import java.math.BigInteger;
-
 public class BodegaDTO {
 
-    private BigInteger id;
+    private Long id;
     private String nombre;
     private String direccion;
+
 }
